Additional optimizations
I applied a few extra optimizations.

First, I added a 'Contact_Flag' to the particle contact method. The idea here is to only go through the effort of updating each particle's contact/friction force if any contact occurs. This should theoretically improve runtime (which my results appear to indicate). A better desciption of this flag can be found in Particle_Contact.cc:56

I also moved a lot of primitive type declarations into for loops (in update_x and update_P). This was done to hopefully improve use of registers. By declaring these varialbes inside of the loop, the compiler is free to discard them after each loop iteration. The hope here is that the variables will be declared in the registers rather than on the stack (and pulled into a register after each iteration). This should improved runtime (which my results also indicate).

With these imrpvoments, I think that my code is running ~10% faster. I was getting runtimes on the order of 110 seconds (for the steel bouncing test) before, after these optimizations I'm getting runtimes on the order of 100 seconds (more like what I was getting before implementing friction). This all implies that the optimizations were worth the effort.