fixed a memory leak
I found (and fixed) a major memory leak in Particle_Contact.cc . I also decreased the density of both the default and stainless steel materials to 10x of their actual values. To accompany this, I decreased the timestep by a factor of 10.

With these modifications, running the first third of a needle insertion simulation takes about 40 minutes. Therefore, running a full simulation will take somewhere around 2 hours. If I were to repeat this with actual densities (3x faster speed of sound) it would take ~6 hours to run a simulation.