Changed IPS to be set per particle array
I am conducting a convergence study on my code. To do this, I'm checking the force output as a function of particle density. To make this work, however, I want to use the same needle for a variety of different blobs. To make this work, I need a different IPS for the body and needle. Therefore, I made the IPS a property of each body, not of the simulation in general (support radius is still a general simulation property). This required making a slight change to Simulation.cc. See Simulation.cc:82 for details.