New convergence study setup
I removed the random partilce strength (set the standard deviation of particle strength to zero in Particle.cc).

I also redid the particle contact so that contact is initiated once the particles are within a specified radius of one another (this is a global variable/not a function of the IPS). See Particle_Contact.cc and Simulation.h:39 for changes.

I also replaced the nice needle with a square punch (this shouldn't depend on particle density). With these changes, I will run a new convergence study. The code is currently set up to run with a 1mm IPS in both the "needle" and the "body"