Fixed materials + did something exceptionally stupid
I realized that the stainless steel material paramaeters were in MPA, not GPA. I fixed this by basically multiplying every constant by 1000 (see Materials.h).

I tried running a simulation with these new materials. This caused a seg fault, which confused me. for this simulation, I had reduced the size of the "needle" to 3x10x3. This turned out to be the problem. After an hour of frustrating debugging, I found that the problem was coming from this. With these dimensions, the size of the needle is smaller than any particle's support radius. This was bypassing the checks in Particle_Neighbor (for finding neighbors in a box), causing the computer to try accessing members that were outside of the Needle array.... I fixed this by changing the size of the needle back to 4x10x4