Included a bunch of stuff + plans to redo code structure
I made a few minor changes throughout the code (in a bid to get it running). I realized, however, that my particle class design is stupid. No class should have that many friends. I thought about it and realized that most of particle's friends are actually body class methods that were made before I came up with the body class. When I wrote the body class, I modified particle's friends to work using bodies rather than arrays of particules, but that was really just a hack of a fix. The real fix is to integrate those friends into the body class. I marked the functions that I believe belong in the body class with "Body method" in particle.h. I will be making particle a friend of body. Expect that change in the next commit (and.... hopefully... working code?).

This change will probably prompt me to eliminate a lot of the particle files (most of the particle helper methods, for example). In order to conserve space, I will be distributing the various body methods over several files (sorta like I'm doing right now with particle).