Made some tweaks to Simulation.cc
I noticed that the first part of the Run_Simulation function can be split off into an independent function. I plan to do a similar thing to set up the boundary data for the first body in the simulation (currently lines 200-300 or so of Simulation.cc). Expect some of those changes in the next commit.