Created new functions to simplify Run_Simulation
I delegated some parts of the Run_Simulation function to new functions. These new functions handle setting the paramaters for the bodies (Startup_Simulation) and setting the BCs for cuboids/needles (Set_Cuboid_BCs/Set_Needle_Bcs). The whole idea is to simplify Run_Simulation.

I plan to redo the indexing variables for Run_Simulation, and to write a function that deals with updating the timers (because what I have right now is terrible).