Redid indexing variables for simulation
The old indexing variables were a mess. I redid them. index variable t is for time (before it was l), b is for bodies (before it was m), and p is for particles (before it was i). I also removed the j,k index variables, as these are no longer needed (they were only used for BC stuff, and BCs are now handled by other functions).

In my next commit, I plan to run a simulation.