Added simulation exceptions + removed Num_Particles simulation member
The Num_Particles member was not being used, so I got rid of it.

I also renamed Body_Needle_Set_Up to just Bodies_Setup.

In my next commit, I plan to make some changes to the way that cuboid are set up. In particular, I want to add a "Cuboid_Parameters" member to the Bodies_Setup stuff. This new member will be a structure. It will house dimensions, as well as BCs for each face.

After that, I want to add something to set BCs for bodies read from file... possibly a function pointer? Idk.