Code runs a simulation!
the code successfully ran a simulation (or, well, It ran part of one without crashing). To make this happen, I had to make a bunch of changes to the "Export_Particle_Positions" method in VTK_File. I also had to make minor changes to "Print_Particle_Forces" in Body.cc (which I renamed to "Export_Particle_Forces").

These changes made me realize that VTK_File.cc should be a part of the Body class. The only important function in VTK_File is "Export_Particle_Positions". This function is analagous to the "Export_Particle_Forces" method already in the Body class. "Export_Particle_Positions" already references tons of members of the Body class. The function implementation would be easier if it were part of the Body class. As such, in my next commit, I plan to move "Export_Particle_Postions" into the Body class. To do this, I will make a new sub-file for the Body class, "Body_IO", which will house all printing methods (including the ones that currently exist and are implemented in Body.cc, as well as "Export_Patticle_Positions").