Redid List class
I am getting ready to redo the way that BCs are applied. My new method involves a list of BC's for each body. The items in this list are "BC" objects. Each BC object will have a component (x, y, or z... this is the component direction  in wich the BC is being applied), a condition (>=, >, =, <, <=) and and an effect (like "set spatial position to 7").

When applying the BCs, cycle through the BC list... for each item in the list, apply the corresponding BC.

There will be separate BCs for Cuboids (for the different faces).

The issue is that this method requires the use of lists. Thus, I needed to update the list class before getting the BC junk to work.

List.h: Renamed Add_Back/Front to Push_Back/Front and Remove_Front/Back to Pop_Front/Back. I also added new methods to get a reference to the Front and Back items in the list. I also deleted the copy construtor and = operators. I also deleted the [] operators (which were slow and not being used). I renamed "Beginning" to "Start".

Update.cc, Neighbors.cc, Damage.cc, Old_Tests.cc: Renamed calls to List members.

Simulation.cc: Formatting.

In the next few commits, I plan to add BC structure (as described above), as well as methods to implement them. I will then make corresponding updates to the simulation stuff. After that, I will add support for a setup file (all the information that is currently in Simulation_Startup, and Simuation.h will be read into a file, rather than being hard set in code... this will allow me to change the simulation without having to recompile... it will also make setup stupidly easy (everything happens in one file)).