Redid BCs
I redid the way that BCs are applied. This process was a little convoluted, but can be summarized as follows: There are two ways to apply BCs. The first is BC structures. These are general and can be applied to any kind of body. The second is the Box_Parameters structure. This structure has 6 components, one for each face of the box. Each face gets its own BC vector. This method only applies to Boxes.

I made a new directory (in the simulation directory) just for BC functions. It's caled Boundary_Conditions.cc.

Right now, all BCs are velocity BCs. In other words, the BCs set the velocity of particular particles. If you want a particular set of particles to not move, you simply define a BC that sets each particle's velocity to 0.

the Apply_BC function works with the Boundary_Condition structure (which is explained in Simulation.h(50-100). They are applied relative to a plane in 3d space (either in the reference or spatial space). This is again exaplined in Simulation.h(50-100).

I made a new structure, Box_Paramters, to deal with Box setup. This strcuture houses the box dimensions, offset, and BCs. The BCs are set as Vectors. the special value Free_BC_Box (which was just chosen as a random value... this isn't great but idk what else to do) designates a free component.

To implement the Box BCs, I modiied the existing "Set_Box_BCs" to work on a general box (before it was set up to only work on the first body in the simulation).