Defined "Array" template class
I made this class to simplify my code and make transitioning Lists into arrays easier. The newly defined Array class has several constructors, including one that whose only argument is a list. The array is generated from this list (which is destroyed in the process). The array otherwise behaves as a standard array. I decided to create this class for two reasons: To make arrays that know their length, and to standardize the way that lists are converted to arrays.

I also modified the "Apply_General_BCs" function to work with an array of Boundary_Conditons as its input

In my next commit, I plan to rewrite several parts of code in terms of the Array class.