Changed "Is_Boundary" member of Body to "Is_Fixed"
The Is_Boundary member was causing me some confusion since boundary conditions mean something else now. Thus, to avoid naming confusion, I decided to rename "Is_Boundary" to "Is_Fixed". I think the new name is a better reflection of what this parameter does (it sets wheather or not the body is fixed in place) and does not cause confusion. Unfortuneatly, making this change was a little more work than I hoped for. I had to rename a getter/setter (Get_Boundary -> Get_Is_Fixed and Set_Boundary -> Set_Is_Fixed... this getter/setter had an awfuly confusing name) and a member in simulation.h (Is_Boundary). All changes in this commit are related to this name change

In the next commit, I plan on finishing the Save/Load stuff that deals with BCs.