Rewrote Load_Simulation using fstream objects + new methods
I rewrote (and tested) Load_Simulation using fstream objects and the new "read_line_after" method. So far, so good. The code is far more readible (woo-hoo). It is undoubetbly slower than what I had before (since reading in a line using the new method requires checking for an entire string rather than just a character). However, sinc this code is only run once (at the start of the program), I don't car. It runs fast enough, so I prefer the new/more readible code to the old code (the difference in runtime will be, at the worst, a few miliseconds).

In the next commit, I plan to write some files to read in the setup file using the same general method as I am now using in Load_Simulation.