Made some function definitions local
I moved a few function prototypes from headerf iles to source files. My goal here is to make the interface between files clear and easy to understand. Part of that means minimizing the number of functions in each header file (so it's very clear which functions should be called externally). I am trying to move the prototypes for all functions that are only called by functions within a source file into the corresponding source file. I already did this for simulation. I now did it for load/save. I was not able to move Load_Particle into Load_Simulation.cc because this function is a friend of the Particle class (which means that the particle class needs to be able to see a prototype).

In the next commit, I plan to clean up the Old_Tests (by wrting a few new tests), and updating the TODO. After that, I plan to implement general BCs in the setup file (for FEB bodies in particular) and write a Read_Me.