Setup file reader is now case-insensitive
Setup_File_Reader.cc: I made the reader case-insensitive. All calls to Contains and read_line_after now pass "false" for the optional Case-sensitive argument. As such, the setup file is now case-insensitive (which sould make writing/working with the setup file a lottttt nicer. You don't need to remember which letters are capitolized and which are not.

Simulation_Setup.cc: removed a line.