I fixed a few typos and renamed some things. Nothing major.
In my next update, I plan to basically totally remove the
"time steps per update" functionality. I think it was a decent
idea, but I also think that it just makes everything a lot more
complicated. Since I don't plan to use it for this project, I see
no reason to keep it.