I made a bunch of changes. I started by trying to redo the "Print net external forces" function in the body class. Let's be honest, that function was useless (who really cares about just the contact and friction forces?). I redid this function to be more useful.
In particular, the function now prints out data on each kind of force
acting on a body (internal, vicious, contact, hg, and friction). I also
calculate the net force being applied to the body.

To make this work, I needed a way to fetch these forces from each
particle in a body. So, I wrote getters for those forces. I also
renamed some of the forces (Force_Int ==> Force_Internal, for example)
to make things more readable.

I also removed the "PARTICLE_MONITOR" variable from the particle class.
This variable originally determined if the particle class kept track of
the Viscous force or not. I decided to always keep track of it (for
debugging purposes... since this code needs a lot of debugging). This
will make things slower. However, since calculating this variable
represents a TINY fraction of the overall runtime, I doubt this will
really impact performance (and the few tests that I've run indicate
that this is indeed the case).

I had to make edits elsewhere in the code to accommodate these changes
(mainly the name changes). However, it now runs.

I have determined that (big surprise), the Hourglass force is causing
all of my problems. Woo-hoo. Now I get to screw with that horrible thing.
I also want to write a "calculate Body Torque" function which basically
does what the "calculate body forces" function does, but for torques
about the body's center of mass. This will hopefully explain why the
bodies tumble. I fully expect that the Hourglass force is causing
these problems... but I'd like to check it anyway.

In addition to debugging, these new output files will make it easy to
look at what's happening to the bodies (from a physics standpoint),
which could make for some interesting insights/make the code more
useful.