I added a option to print out the torques acting on each body. I added this as an option to the setup file (right after the "print body forces" option). If set to true, then the code generates a file which is very similar to the net Force files (for a body). The only difference is that we calculate torques, not forces.
To make this work, I had to write a function to calculate cross products
I added this as an auxiliary function to the vector class (See vector.h
and vector.cc). This just calculates the cross product component-by-
component.

The function that actually prints out the torques is called
"Export_Body_Torques" and is located in the body class (its definition
is in Body_IO.cc).

This function is almost identical to the "Export_Body_Forces" function
(in the same file). However, since this calculates torques, I first
need to calculate the body's centroid. I calculate the net torque on the
body due to internal forces, contact forces, friction forces, viscous
forces, and Hour-glassing forces.

The code appears to work (the torques it gives me agree with the
way that the body is rotating). Unsurprisingly, the Hour-glassing force
is the biggest contributor. Somewhat surprisingly, however, is that the
internal force seems to also play a big role. What's more, the internal
force imparts a torque on the body which acts in the same direction as
the torque due to the hour-glassing force. The two torques work
together. This is probably why the bodies tend to tumble in my code.

This information is quite useful. Hopefully it will allow me to fix
these bugs (somehow).... possibly by playing with the HG constants.

I also removed bin/SPH from the set of files being tracked by git. I
plan to add flags to the makefile which compile the code for the local
architecture. As such, the executable will ONLY run on the machine that
compiles it.

As I was writing this, I realized that there are no longer any files
in the bin directory that git is tracking. Oops. I expect that, in this
commit, the obj directory will disappear. Oops. I will add a stupid file
(.gitkeep) to the obj directory in the next commit so that git keeps
that directory in the repo. Oops.