When I looked closer at the torque data, I noticed that the torque due to the viscosity was almost identical to the torque due to the internal forces. That made me remember that the viscosity force is basically added into the internal force (when calculating P in update_P). Since my code now tracks the viscosity tensor and force at all times (before it only did so when the Particle monitor was defined), I thought the code would be better if kept the viscosity term separate. So... I did just that (see update_P, update_x).
This change let me easily tell that the viscosity (not the real
internal forces) are causing the tumbling (along with the hour
glassing stuff).

I am a little worried about this change. The viscosity is supposed
to be part of the constitutive law... not a true separate term.
I may add it back in (go back to what I used to have), and just
separate it out in the function that calculates the forces and torques.

Anyway, this change let me see what was going on in my code. As it
turns out, the two corrective terms are causing all my problems.
Woo-hoo. Lots of stuff to tune, I guess (the constants that control
these corrective terms). At least I know (with confidence) where the
problems are coming from.