I changed the code a bit.
I didn't like the fact that my last commit modified the way that P
was calculated. I want P to be calculated the way that the constitutive
law says that it should be calculated (with the viscosity and elastic
terms). Thus, I undid the changes that I made to update_P and update_x
in the last commit.

I still wanted a way to identify which parts of the force were due to
elastic and viscous terms. However, I realized that the internal force
is the sum of the elastic and viscous forces. Since I already know the
internal and viscosity torques, I can find the elastic torque by taking
the difference of the two. This is exactly what I did (see Body_IO.cc)

I also renamed "Internal Torque" and "Internal Force" to "Elastic Torque"
and "Elastic Force", respectively. With these changes, the output body
forces and torques are pretty informative.

I hope to use these files to tune the HG and viscous terms, thereby
improving the stability (and correctness) of my code.