Contact.cc: I updated a few comments + added a barrier to the end of the contact function. I realized that I needed this barrier (idk why I thought that I didn't... perhaps in an older version of the code, it wasn't necessary... but it is now!).
Update.cc: Fixed a few comments.

Simulation.cc: Renamed t as time_step (I think this name is more
explicit). I Also updated a few comments.

In the next commit, I plan to begin rewriting the contact algorithm
using a "buckets" approach. I expect this to take some time to finish,
however, since the "buckets" approach will entail 8 nested loops,
a lot of steps, and some interesting logic.... it promises to be
a lot faster than the existing contact algorithm, however (
especially when there are a lot of particles and/or bodies).