I rewrote a large part of the contact algorithm (again). I was not happy with the performance I was getting. The new algorithm was running slower than the old one (granted, this particular simulation runs reasonably well on the old algorithm. On different simulations, its very likely that the new one would have destroyed the old. still, however, this was not good). I thought about the code a bit. The new algorithm basically consists of two parts: Sorting particles into buckets, and applying the contact algorithm on a bucket-by-bucket basis. The sorting part was happening totally sequentially. Each thread did the same computations /generated the same buckets simultaneously (talk about redundant calculations). I did it this way because openMP is dumb and doesn't let you make shared variables within a parallel region, so it was going to be tricky to make this work in parallel.
Further, since this part used linked lists, with a node for each
particle in the bodies, this part was using a LOTTTTTT of mallocs.
Enough to tank performance. I decided to try rewriting the code to
eliminate both of these issues.

To do that, I first used a global shared buckets array (and some
shared buffers) instead of thread specific buckets. Global variables
are shared, so this was sorta a hacky way to get shared variables
where I wanted them. I made these variables static because they should
only be used within the Contact.cc file.

Any who, I then rewrote the "sort into buckets" part of the contact
algorithm. The rewrite does this step in parallel. Unfortunately,
making this work requires a lottttt of synchronization and a lot of
lines of code. Making this change added 200 lines. Awkward. Amazingly,
it appears to have worked. Even better, performance is better now...
a lot better. Like better than in the old algorithm. This is a great
sign.

I realized that the buckets are tiny (like 1mm)... which may be too
small (too many empty buckets to get good performance). They're small
because the contact distance is quite small. In future commits, I plan
to try larger buckets. Larger buckets mean more computations per bucket
In fact, if we double the bucket size then we should get 64 times more
computations per bucket, even though we only reduce the number of
buckets by a factor of 8... this increases the number of computations
by a factor of 8. Or, well, it does if mallocs are free and setting up
more buckets is free. Clearly, that isn't true. Thus, since most
buckets are empty, I think that slightly larger buckets may reduce
runtime by eliminating a lot of mallocs (one malloc per bucket to
get Array_A and Array_B) and setup, even though it will mean more
computation per bucket. Idk, it's a variable to play with at the very
least.