I fixed a bug (actually several) in the contact algorithm + fixed numerous typos throughout the code.
There were two big problems in the contact algorithm. The first
had to do with calculating the sub_cuboid (now called a cell) size.
The z dimension of a sub_cuboid (cell) we calculated as (y_max - z_min)
/ Contact_Distance. Oops. I fixed the "y" to a "z".

So why did the code run before? Because in the simulation that I
happened to be testing, y_max was much bigger than z_max. As such,
all that this typo did was make the buckets way too large in the z
direction. This is okay, because larger buckets doesn't crash the code.
All it does is make the code slightly less efficient.

This change, however, revealed another bug! Basically, the way that
the algorithm works, if a particle lies on the boundary of two cells,
the algorithm assigns them to the cell with the bigger index. This is
problematic if the boundary corresponds to the cell with the largest
x, y, or z coordinate. In these cases, the algorithm tried to assign
the particles to a cell that did not exist (one with a larger x, y, or
z) coordinate than is allowed. I fixed this by throwing in some checks.
See line 230 or so of the new code... I explain this in detail.

Aside from these changes, I mainly just fixed silly typos in comments
throughout the code.