I fixed several small bugs and improved a few printouts.
Body.cc: This had a legitimate bug. The "Set Box Dimensions"
function set each particle's ID... or, that's what it was supposed
to do. The issue was that I incremented the index variable BEFORE
setting the particle's IDs. This meant that particle 0 never got its
ID set. It also meant that the code tried to set the ID of a particle
with ID num_Particles (which doesn't exist!). Somehow, this didn't cause
a segmentation fault. Thankfully, valgrind found it (Remember, I haven't
been able to use valgrind on my mac because mac has problems).

Body_IO.cc: I improved the printout. I also changed the code so that
it only prints out tau if it is a damageable body. IF the body is not
damageable, then this variable is uninitialized/calling it prints an
undefined value.

Damage.cc: No major changes.

IO_Ops.cc: Fixed a typo. This function tried to include IO_OPS.h, a
file which does not exist... somehow, this compiled correctly? I
guess that g++ ignores case when including? or something? not sure.
I honestly don't know how this code compiled with this typo. I only
caught it because I tried to compile the code on a GCP machine, and
that one threw an error.

List.h: Same bug as in OP_Ops, but with "classes.h" (there is no file
called "classes.h", the correct name is "Classes.h"). Again, i have
no idea why the code compiled with this typo.

Setup_File_Reader.cc: I greatly improved the printout part of this
file. It's much easier to read now/better formatted/prints true and
false instead of 1 and 0.

Simulation_Setup.cc: No major changes.