I wrote a new Body IO function: Export_Box_Boundary_Forces. This function is used to pint the net forces acting on each of the six boundaries of all box type bodies in a simulation (which are not fixed). I plan to use this function to analyze the forces applied to bodies in various simulations (such as stretching, shearing, etc).
To implement this function, I had to make a bunch of changes. Throughout
the code and add a new option to the setup file.

I also wrote a Body exception "Not a box" which is thrown whenever a
non-box body tries to do something that only boxes can do. I haven't
fully implemented it, but plan to do so soon.