I moved the timing tests into the tests directory and added function descriptions.
I have no idea why the timing tests (which I haven't used in ages)
weren't in the test directory... that directory is not only for
unit testing! Oh well, it's fixed now.

I also added function descriptions throughout the code. A few functions
were lacking descriptions... so I fixed that.

In my next commit, I plan to revamp the Operation_Count function. Rather
than counting the number of times that a function is called (something
that gprof does), I want it to keep track of the number of flops. For
this to work, I will have the function keep track of counters for
additions, subtractions, multiplications, divisions, etc. In each
function (or, at least those which are used for the simulation...
i.e., not IO functions), I will count (by hand) the number (and type)
of flops performed by that function. I will then increment these in
the global counters (if the counting flag is defined). I will need to
make sure that these bits are thread safe.