I revamped the Operation Count namespace.
It counts flops (categorized by the type of operation used to increment
the flop) instead of function calls (which gprof already does!).

I intend to use this to basically determine the performance of the code.
Unfortunately, for it to work, I need to implement operation tracking
in every function in the simulation. Currently, I've implemented it for
the vector and tensor classes. I've also implemented it for the two
functions I wrote in update_x which eliminate operator overloading.
There are, however, many more flops to count.

Currently, the operation count stuff does NOT run in parallel (it's
just asking for data races). I want to fix this with atomics or
something... idk... we'll see. There's a lot of work to be done before
the operation count thing will be totally useful.

I also tweaked a few things in Simulation::Run. Idk what's wrong, but
I'm pretty sure that the "print runtime" is wayyyy off when the code
runs in parallel. I want to get to the bottom of it.