I expanded Operation_Count to functions in the Body and Simulation directories. I also tweaked the compiler flags and fixed a bug with timing the time to print to file.
Body.cc, Contact.cc, Update.cc: Added operation count stuff to these
functions. I choose to not add it to Damage and Neighbors because
the former is called infrequently and because Jennifer is rewriting
parts of the latter.

Simulation.cc: The code used to give weird timing numbers for "Time to
pint to file" when run in parallel. I finally figured out the issue.
On most time steps, the condition time_steps%time_steps_between_prints
would evaluate false. As such, the code would immediately jump to the
bit of code that evaluated the time for the printing bit. In principle,
this wouldn't cause a problem (it would just log a zero). However, as
far as I can tell, the "Get_Time" function runs slower than the
"Time_Since" one. As such, the following situation can cause a bug:
Two threads finish a time step. One of them begins calculating the
time to update_x (and resetting the time). The other moves onto the
next time step, the condition time_steps%time_steps_between_prints
evaluates false, so the thread starts computing Time_Since (to
update the Print_Time). At this stage, the first thread has called
Get_Time, but hasn't updated. Therefore, time2 is set to the time
before update_x finished, not after it. This only happens when
one thread tries to execute Print_Time += Time_Since(time2) before
another thread has finished running Get_Time (at the end of the
previous cycle). In other words, we have a data race on our hands.

To fix this, I did a few things. First, I moved the timing bit
inside of the conditional, so that it only updates the Print time
if there's something to print. This should give the thread that's
updating Get_Time to finish before the time get's updated.
Unfortunately, this change created another bug. The print time
appeared to go to .0001, far lower than I know it is (about .04,
this is the runtime when executed on one core... multiple cores
shouldn't change the runtime of this 1 body simulation).
The bit of the code that printed to file used an orphaned for loop
(in Export Bodies data). This for loop was over the number of bodies.
As such, most threads just skipped this loop and went straight to
the Print_Time line. To fix this, I decided to just put the whole
"print out runtime, print to file, etc.." bit inside of a single
region. I also changed the Export_Bodies_Data function by removing
the orphaned parallel for loop. This function is no longer thread
safe as a result, but I'm okay with that. With these changes, the
code appears to give correct timing data. Even when run in parallel,
Print_Time ends up around .04, exactly where it should be.

In the next commit, I want to try making the Operation_Count stuff
thread safe.