I changed the OP_Count to work in parallel.
I basically added a ton of atomic update statements to guarantee that
everything happens sequentially (except for in contact, which has
computations that every thread does... I used a single construct for
those). Unsurprisingly, this change tanks performance (like seriously
tanks it). Weirdly, the code runs slower in parallel than in serial. Oh
well. It's not supposed to run fast in this mode.

It is concerning, however, that the OP counts are different in serial
and parallel. These differences are not due to randomness (I tried
making the particle strength SD to zero, this didn't eliminate the
discrepancy). Some of the OP counts are the same (division, log, and
Sqrt). I believe that the eigenvalues function (in the tensor class)
is causing the issues. This function has a conditional which exists
only because floating point errors exist. The OP count in these
branches corresponds exactly to the operations that are giving different
results. Thus, as far as I can tell, the difference exists because
when the code runs in parallel, the number of times that we get out of
bounds results when computing the eigenvalues is different from when
we run it in serial. I don't know why, but that's my best guess.

Either way, the discrepancy is small. As far as I can tell, I have
made the operation count stuff run in parallel, which was my goal.