Fixed a printing bug + improved parallel OP_Count consistency
I went down a big rabbit hole to try and fix the OP_Count. Along the
way, I found a bug in the Simulation code. After the time steps
finished, every thread called Export_Bodies_Data. This should have
only happened once. I moved the function inside of a single construct.

After I disabled all OP_Counting on the eigenvalues function, I noticed
that there were 2058 too many division, 6174 too many additions,
and 12,348 too many multiplications. These numbers form a ratio of 1:3:6.
Also, 2058 is double the number of particles (1029 = 7*21*7) in the
simulation I was running. I assumed that I was doing something wrong
twice per particle which happened to add 2 divisions, 6 additions, and
12 multiplications. This assumption was wrong. Instead, I found that
the problem lay in the Export_Bodies_Data thing. The simulation I was
running had Boundary Force printing turned on. In that function, I
found that the computation that calculated the net force on each
face happened to use 1 division, 3 additions, and 6 multiplications
per boundary particle. There are 21*7*4 + 7*7*2 = 686 boundary
particles in the simulation I was running. This code was called 3 times
too many (4 threads, this function was called 4 times at the end of
the time steps when it should have been called just once, so 3 extra
calls). 686*3 = 2058! woo-hoo. I found the bug. Implementing the
change above (move the final call to Export_Bodies_Data into a
single construct) eliminated the discrepancy in the 4 core and
single core OP counts (when counts for calls to eigenvalue were
disabled).

Body_IO.cc: Along the way, I discovered that there was a division per
boundary  particle that wasn't being tracked. I decided to move that
into  an operator overloading bit in (thus the 1000 ==> 1000000). This
also has the benefit of incurring fewer divisions (which should improve
runtime ever so slightly). This accounts for the changes in this file.
I explained this change in a comment on line 295.

After I discovered the bug above, I re-enabled OP_Counting in the
Eigenvalues function. Not surprisingly, this caused discrepancies to
reappear. As explained in the last commit, I think this is because
the eigenvalue function runs different branches depending on floating
point errors. Floating point errors will be different when stuff is
done in parallel than when it is done in serial. As such, we expect
slight differences in the OP_Counts due to the eigenvalues function.
However, I can now say with confidence that the discrepancies are
entirely due to this randomness (in the Eigenvalues function/fp errors),
which is what I wanted to get to.

Contact.cc: I realized that I needed to add atomic directives to
the single stuff... oops.

Body.cc: I decided to remove the Op_Count stuff from the setup functions
Set_Material and Set_Support_Radius. These functions are used in
setup, so the FLOPs in them should not be counted (because timing stuff
only looks at the things run in parallel, which does not include setup)