Implemented the algorithm used to sort particles into various buckets (which is already used in Contact.cc) in Neighbor.cc.
The code still needs to be debugged. It appears that the neighbor list isn't being filled for some particles when they definitely have neighbors.

Simulation_Setup.cc is changed to use this new neighbor algorithm instead of the one for Boxes, so that we may debug the code on our sample box simulation.