I fixed a few bugs with the bit of code that deals with loading and saving simulations.
Bug 1: "Load_Simulation.cc" attempted to set up a body's neighbors
even if that body was fixed in place. This was a bug. I fixed it...
see line 305 in Load_Simulation.cc.

Bug 2: The Load_Simulation function tried to set neighbors even for
damaged particles. I added a line to skip these. See line 309.

Bug 3: The Load_Simulation function didn't have a way to handle
singular tensor exceptions. It is theoretically possible for a particle
to have an invertible A tensor in one simulation, but not when
we load that simulation. This is because the saving process loses
precision (we only save a few decimal places). This change may cause
a particle's A tensor to become singular, in which case, we need to
damage the particle. I added a bit of code to handle this possibility.
See lines 300 to 330.

Bug 4: The Save Simulation code did not save "Gravity Enabled".... so
all loaded bodies defaulted to "Gravity Enabled = false"... oops. I
fixed this. See line 121 of Save_Simulation.cc.

Bug 5: In Simulation.cc, the code used to print particle positions
at the first time step. This is problematic if we're running a
loaded simulation, however, as that data would have been printed at
the end of the previous simulation... therefore, I added a bit of code
which skips the print if loading from sim and time step = 0. See
Simulation.cc lines 68 to 85.

I also redid the Load Monitor stuff to be more readable. I also renamed
"Save Data To File" to "Save Simulation" in the setup file... which is
a much better name. Finally,